+++ NEWS.md
@@ -2,13 +2,16 @@
  NEWS for R Package "medfate"
 -------------------------------
 
+# Version 2.7.2
+- SpParamsMED include the results of parameter estimation exercises
+- RGRsapwoodmax and RGRcambiummax regulate sapwood formation for shrubs/trees, respectively 
+
 # Version 2.7.1
 - Revision of recruitment model, with the addition of a recruitment probability (ProbRecr) within the bioclimatic limits
 - Temperature effects on sapwood conversion to heartwood
 - Correction of estimation of Psi_Extract from turgor loss point
 - Estimation of RSSG from shade tolerance
 - Minimum DBH parameter in stand metrics
-- SpParamsMED include the results of parameter estimation exercises
 
 # Version 2.7.0
 - Basic water balance model with relative water content